cdm-agent/requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# MCP dependencies
mcp>=1.0.0

# Database dependencies
psycopg2-binary>=2.9.9

# Environment and configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

